feat: Split logic for Main Store
